Move use-shopping-cart-manager to lib/shopping-cart
