+++ client/lib/shopping-cart/use-shopping-cart-manager.ts
